move rootchain code to different files
